Refactor: 컴포넌트 폴더 하위 폴더 생성 및 위치 변경
