+++ src/components/domain/add/steps/step1/Step1.tsx
@@ -3,7 +3,7 @@
 import { Search } from 'lucide-react';
 import { useState } from 'react';
 import { Button, Input, Txt } from '@/components/atoms';
-import AddressSearchModal from '@/components/domain/add/AdressSearchModal';
+import AddressSearchModal from '@/components/domain/add/steps/step1/AdressSearchModal';
 import { SearchButton } from './SearchButton';
 
 export default function AddAdress() {
